RELATIONSHIPS AND ROUTE PROTECTION,  CUSTOM VIDEO PLAYER

+++ controllers/videoController.js
@@ -55,8 +55,11 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
   });
   //console.log(newVideo);
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videosDetail(newVideo.id));
 };
 
@@ -68,8 +71,7 @@ export const videosDetail = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id);
-    //console.log(video);
+    const video = await Video.findById(id).populate("creator");
     res.render("videosDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -81,8 +83,12 @@ export const getEditVideo = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    const video = await Video.findById(id).populate("creator");
+    if (req.user.id === video.creator.id) {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    } else {
+      res.redirect(routes.home);
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -107,7 +113,12 @@ export const deleteVideo = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id).populate("creator");
+    if (req.user.id === video.creator.id) {
+      await Video.findOneAndRemove({ _id: id });
+    } else {
+      res.redirect(routes.home);
+    }
   } catch (error) {
     console.log(error);
   }
